Given a Binary Search Tree and a target number, return true if there exist two elements in the BST such that their sum is equal to the given target.
思路：
This method also works for those who are not BSTs. The idea is to use a hashtable to save the values of the nodes in the BST. Each time when we insert the value of a new node into the hashtable, we check if the hashtable contains k - node.val.